* added v3 folder with standalone rails 6 code for the v3 api server.  needs to be dockerized...

* make up to 3 attempts to generate unique ip + token

* fix migrations for indexes and foreign keys so they don't crash when running multiple times

* roles & permissions

* user_controller tests

* added get_roles

* users and permissions tests

* fix generic check token call when not validating a specific role/permission

* migration files for roles tables

* initial without previous git info

* Refactor: use sessionStorage, Add API helpers, Check for valid token on /login

* Refactor validate token to async with boolean return

* Refactor styling

* Add divider

* Refactor auth, add fragment to render header on every page

* Update Cypress tests

* Add eslint with Airbnb styling guide, clean up styling errors

* changed user_controller to token_controller and changed api calls. created a 'check_token_for_permission()' method to gate each api call

* changed is_role? to has_role?

* guards instead of chained if statements on has_role?

* changed 'ins' to 'conditions'

* Fix styling issues

* Add cypress screenshots and videos to gitignore

* change 'role' to 'permission'

* seed.rb for test db

* Add .nyc_output to .gitignore

* Refactor LoginDialog POST request  to func in API

* docker

* remove references to v3. TODO - fix eula

* notes re: retired and admin permissions

* Remove unused Login Button
 Refactror UserMenu to use helper api functions

Remove auth folder

* Add hamburger menu component

* Add prop-types package

* Update import paths
Replace menu loop with LeftHamburgerMenu component

* Add placeholder pages and tests for all pages in hamburger menu

* rubocop app -a

* rubocop app -A

* more rubocop

* Refactor component naming

* change port 3000 to port 3001

* docker frontend + backend

* docker-compose fix for /app/node_modules (now react pages update on the fly)

* Update cypress tests

* Fix spelling

* Revert global replace in legacy code

* Add folders &  update tests

* Refactor axios requests to handle special characters.

* fix API.js  linting errors

* use header statuses - TODO: update docs

* Update Spacing

* Refactor to handle status headers

* Resolve conflict

* Add New button to sample types page
Add sample type form
Link New button to New sample type route
Resolves #453

* Add container and styling

* Add Description input

* Udpate tests & form

* Add btn WIP

* WIP sample fields form

* WIP fields

* WIP -- fix alignment

* WIP

* WIP: add close icon

* Delete button placement and choices width

* Handle fields form state in parent component
Add remove field function
Typecheck with prop types in fields form

* Add cypress-react-unit-test dependency

* Cypress unit test running

* WIP fields unit testing

* Replace grids with fragments

* Spy on change event

* Update Cypress

* Test Field Type Input

* wip required tests

* More testing

* Refactor per design convo w Sean, update tests wip

* WIP connecting backend

* Display list and first sample type

* disable react/jsx-one-expression-per-line

* Add getTypeById

* Create shared standard modal component

* Create shared loading backdrop component

* Create shared Button components

* Create FieldTypesList

* Create ObjectTypesList

* reafactor naming, dark bol, add testName

* Use LinkButton

* Use StandardButton, Loading Backdrop, WIP aft

* Refactor use StandardButton & more SFCs

* Use LoadingBackdrop, LinkButton & SideBar SFCs

* create Show component

* create SideBar SFC

* refactor onClick and naming

* Add create sample type

* Clean up

* save form

* clean up

* refactor component tests

* Delete button test

* Buttons tests

* Render app bar as nav

* Add location to redirect

* add back button test

* Remove responsive layout

* Handle edits

* Add edit function

* Add grid layout, extract styling

* Remove grid from side bar

* Update styling

* fix proptypes

* Fix checkboxes

* Height and alerts

* Cleanup, styling, disable submit

* Refactor fields into folder, add shared utils file

* Cleanup and design changes per Sean

* Update App.jsx

fix lint error

* Merge branch 'v3-refactor' into sample_types

* Refactor API structure, Alerts, comments

* new sample type - disable save button

* .eslintignore for dev

* backend

* get sample types - include description for first sample type

* sample type form - only call setisloading once

* object_types

* setup

* object_types

* backend cleanup

* setisloading wip

* update propTypes

* Updates from PR, add & refactor testing

* Replace sessionStorage with localStorage
Add app min width

* refactor for isLoading

* Fix test

* fix for null field name when editing sample type (and adding a new field type)

* Fix merge issues

* move setalertprops to app.jsx +  pass alert as sessionstorage when delete object type

* Disable edit button

* Refactor cyrpess env variables

* New Component

* update eslint

* execapis wip

* Update schema

* Delete .env_files directory

.env_files should be untracked

* Delete data/db/development directory

data files should not be tracked

* eslint

* Delete data directory

Contents of data directory should not be tracked

* Clean up

* Add jquery
Update eslint rules

* Update schema

* group spec

* wizard spec

* object types spec

* change to preferences api call + support for wizards by first letter

* new_samples_private check for '1' instead of 1

* merge up-to-date backend code

* changed the wait to 3 sec (it was brittle at 1 sec), but there has to be a better solution

* extended user info on login and verify token

* added waits on cy tests to eliminate timing issues. there has to be a better way to solve this, though

* PR changes
Testing

* user spec - removed check for 200 and  replaced with expected info on screen + added tests for reset button

* group spec - remove check for 200/201 response code

* wizards spec & object_types spec - remove check for 200/201 response code

* Testing

* Update file name
Do not show add option button when no samples to add

* remove cypress/react component testing, styling cleanup

* mariko's comments on the PR

* edit baseline logic and comments re: handling session timeout

* Styling & package updates

* Styling

* Fix imports

* Update styling and packages

* Fix tests

* change user

* run migration

* Remove unused test files & tests

* update styling

* Update eslint rules

* remove eslint prettier for now

* eslint everything except sample types

* eslint sample types

* remove references to '6' for 'retired'

+++ frontend/src/helpers/api/axiosInstance.js
@@ -40,7 +40,8 @@ axiosInstance.interceptors.response.use(
           break;
         case 401:
           if (window.location.pathname !== '/login') {
-            if (currentSessionToken) {
+            // Delete the token if the error message is "Session timeout"
+            if (currentSessionToken && data.error === 'Session timeout') {
               axios.post('/token/delete');
               localStorage.clear('token');
             }
